Move st.data_editor tests into st_data_editor_editing_test.py
Relocate all st.data_editor-specific e2e tests (row add/delete, cell
editing, edit-state persistence, toolbar, and key-based CSS class) out of
st_dataframe_interactions_test.py so that file now covers only the
read-only st.dataframe element. Moved tests use key-based locators against
new keyed editors in st_data_editor_editing.py, and the corresponding
snapshots are relocated accordingly.
